Implement Range

+++ src/main/java/com/helmo/archi/google/googleuse/service/BirdService.java
@@ -1,7 +1,9 @@
 package com.helmo.archi.google.googleuse.service;
 
 import com.helmo.archi.google.googleuse.model.Bird;
+import com.helmo.archi.google.googleuse.model.Observation;
 import com.helmo.archi.google.googleuse.repository.BirdRepository;
+import com.helmo.archi.google.googleuse.repository.ObservationRepository;
 import org.springframework.core.env.Environment;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Criteria;
@@ -12,22 +14,26 @@ import java.util.Arrays;
 import java.util.List;
 
 @Service
-public class BirdService implements BasicService<Bird, Long> {
+public class BirdService implements AccessRange<Bird, Long> {
 	
 	private final BirdRepository brdRepo;
+	private final ObservationRepository obsRepo;
+	
 	private final NextSequenceService nextSeq;
-	private final MongoTemplate monqoTemplate;
-	private final Environment env;
+	private final MongoTemplate mongoTemplate;
 	
 	private final String COLLECTION_NAME;
+	private final Bird DEFAULT_BIRD;
 	
-	public BirdService(BirdRepository brdRepo, NextSequenceService nextSeq,
-	                   MongoTemplate monqoTemplate, Environment env) {
+	public BirdService(BirdRepository brdRepo, ObservationRepository obsSrv, NextSequenceService nextSeq,
+	                   MongoTemplate mongoTemplate, Environment env) {
 		this.brdRepo = brdRepo;
+		this.obsRepo = obsSrv;
 		this.nextSeq = nextSeq;
-		this.monqoTemplate = monqoTemplate;
-		this.env = env;
+		this.mongoTemplate = mongoTemplate;
 		this.COLLECTION_NAME = env.getProperty("mongodb.birds");
+		
+		DEFAULT_BIRD = brdRepo.findOne(0L);
 	}
 	
 	@Override
@@ -52,6 +58,7 @@ public class BirdService implements BasicService<Bird, Long> {
 	}
 	
 	public Bird update(Bird toUpdate) {
+		if (toUpdate.equals(DEFAULT_BIRD)) throw new IllegalArgumentException("Can't update this bird");
 		Bird bird = brdRepo.findOne(toUpdate.getId());
 		
 		bird.setName(toUpdate.getName() != null
@@ -73,17 +80,36 @@ public class BirdService implements BasicService<Bird, Long> {
 	
 	@Override
 	public void deleteById(Long id) {
+		if(id == 0) throw new IllegalArgumentException("Can't delete this bird");
+		manageObservationWhenDelete(id);
 		brdRepo.delete(id);
 	}
 	
 	@Override
 	public void delete(Bird... birds) {
-		brdRepo.delete(Arrays.asList(birds));
+		List<Bird> arraysBird = Arrays.asList(birds);
+		if(arraysBird.contains(DEFAULT_BIRD)) throw new IllegalArgumentException("Can't delete this bird");
+		Arrays.asList(birds).forEach(this::manageObservationWhenDelete);
+		brdRepo.delete(arraysBird);
 	}
 	
 	@Override
 	public void delete(Bird bird) {
+		if(bird.equals(DEFAULT_BIRD)) throw new IllegalArgumentException("Can't delete this bird");
+		manageObservationWhenDelete(bird);
 		brdRepo.delete(bird);
+	}
+	
+	private void manageObservationWhenDelete(long id) {
+		List<Observation> observations = obsRepo.getByBirdId(id);
+		observations.forEach(
+			  o -> o.setBird(DEFAULT_BIRD)
+		);
+		obsRepo.save(observations);
+	}
+	
+	private void manageObservationWhenDelete(Bird bird) {
+		manageObservationWhenDelete(bird.getId());
 	}
 	
 	public List<Bird> findSingleByArgs(String key, String item) {
@@ -101,7 +127,7 @@ public class BirdService implements BasicService<Bird, Long> {
 	public List<Bird> findAllByArgs(String key, String args) {
 		Query findQuery = new Query();
 		findQuery.addCriteria(Criteria.where("data." + key).is(args));
-		return monqoTemplate.find(findQuery, Bird.class);
+		return mongoTemplate.find(findQuery, Bird.class);
 	}
 	
 	public List<Bird> findAllByArgs(String key, double args) {
@@ -109,7 +135,7 @@ public class BirdService implements BasicService<Bird, Long> {
 		findQuery.addCriteria(Criteria.where("data." + key)
 			  .gte(args - args * 0.5)
 			  .lt(args + args * 1.5));
-		return monqoTemplate.find(findQuery, Bird.class);
+		return mongoTemplate.find(findQuery, Bird.class);
 	}
 	
 	public boolean exist(long id) {
